fix: Dont use region subdomain api origins (#67)
The comment in the code also says it... but the /iframe/ view is setting
our cookie into an origin that includes the org-subdomain. so like
`document.cookie = 'sid=123; domain=acme.sentry.io`

Therefore we don't have auth when we try to use a region-subdomain like
`us.sentry.io` for the api calls.
The result is slower api calls because the server will need to sort out
which region to use, and jump over to it. We can fix this in the future
though.