feat(icons): Import platforms icons and load them up into a story (#99)
Getting SVG Icons loading has been tough with the way `platformicons`
package is setup, and the build system here.

So instead of doing what's potentially the 'rigth' thing which could be
to adjust how platformicons is built (and risk breaking sentry &
sentry-docs), I pulled in `vite-plugin-svgr` and loaded up all the svg's
directly. This is probably fine because we need all those files in the
bundle, or else we need them hosted somewhere to reference on demand.
Doing the on-demand thing too might mean setting up a cdn.. or maybe
something like loading the svg directly from https://esm.sh/ would work
too....

for now this is working so i'm happy, but it's up for iteration.