feat: Render Platform Icons in the Issue/Feedback panel titles (#115)
There's a little `isLoading` state, then the icon will pop in for the
project that the toolbar is configured against:

<img width="337" alt="SCR-20241126-lxfo"
src="https://github.com/user-attachments/assets/6c176288-92dc-480f-8646-3943ebeaf28b">

After the project data is loaded once, it's available for the other
panels and places where we need it, so in those cases we don't see the
loading state. `useQuery` caches results and de-duplicates requests for
us.