feat: RelativeDateTime auto-updates the relative rendered string (#117)
Just have it auto-update the string `24min ago` to be `25min ago` if
you've looking at the page for a while.

The update interval can be set manually to either `second`, `minute` or
`hour` but if you don't set it then we'll see how old is the date, and
set an interval that's appropriate. If the timestamp starts being really
recent, then we'll update it every second, but if you keep the page open
for hours and hour then we'll get a new value from
getRecommendedInterval()` and slow down the updates because it won't be
changing as often anyway.

I'm using `dangerouslySetInnerHTML` here to keep react away from the
changing values. If i were mutating the children directly and react
wanted to re-render it could be surprised that the value is not what it
expected.