ref: Refactor locale definition so its statically setup at app startup (#118)
This moves the `formatDistance` definition out into it's own file so we
can define it once and import it whenever; instead of each component
creating it's own instance of the function.

I also iterated on the `suffix={string}` param and converted that into a
`locale` param. With `suffix` and/or `prefix` we endup with
hard-to-translate strings in the end, which is a bad pattern to setup.
Instead we can pass in a `Locale` object that can have custom methods to
format the strings. So far i've defined two locale's with different
`formatDistance` implementations: one implementation adds the `"...
ago"` suffix, and the other adds the `"... old"` suffix. Pass in the
desired locale as-needed.

This type of locale system could be extended to other formatters, like
number, currency and full dates too. I like it so far!