feat: Iterate on ApiProxy to keep login/logout message off the MessageChannel (#142)
Related to: https://github.com/getsentry/sentry/pull/81942

Changes:
- The `request-login` and `request-logout` messages are going over
`iframe.contentWindow.postMessage()` now. The MessagePort is reserved
for when the iframe is logged in and properly configured.
- When state _changes_ we dispose of any ports (would only been
connected if we were previously logged in) and reload the iframe. This
ensures that we get a new state told to us from the html file
- We've dropped the `sentryApiPath` config value, because the config
endpoint (domain, region if needed, and this prefix) is set inside the
iframe js scripts. The `fetch` command of the proxy can only make fetch
requests to the sentry api now, not to other domains.