ref: Replace framer-motion with Transition from headlessui (#156)
https://headlessui.com/react/transition

Using `<Transition>` made it easier to have a nice fade in/out animation
compared to the size change thing we were trying to do before. Also,
it's adding much less to the bundle size.

Bundle size numbers:
- framermotion added 107kb to the bundle
- i tried something called
[auto-animate](https://auto-animate.formkit.com/#plugins), which added
7kb
- this `<Transition>` component added 6kb

While bundle size is not a primary concern for this dev focused build,
it's a nice win.