feat(flags): Create an initial, basic, OpenFeature adapter to read feature flags (#194)
In this PR i've added a makefile and config to be able to run
https://flagd.dev/ doe dev purposes, in order to feed back into the the
OpenFeature SDK and then through the adapter into the UI, and a new
Adapter to consume those flags!

The OpenFeature adapter itself is kind of janky, there are no public
methods in the OpenFeature SKD for getting 'all flag' info back out. The
OpenFeature type interface is something like:
```
type Provider = {
  evaluateFlag(...): Result
};
class OpenFeature {
  constructor(provider: Provider)
  evaluateFlag(name, default) { return this.provider.evaluateFlag(...); }
}
```
And each provider implements their own 'private' cache of flag data, in
different formats, in order to evaluate it. What's beneficial to us
though it that most client side providers use the [Static-Context
Paradigm](https://openfeature.dev/specification/glossary/#static-context-paradigm)
to load data.

So this adapter is probably the most jank out of all because it's got
this Provider abstraction and lots of private data. Some providers, like
the
[launchdarkly](https://github.com/open-feature/js-sdk-contrib/tree/main/libs/providers/launchdarkly-client)
one seem to defer to the standard launch darkly sdk. So we could target
that independently, and have this OpenFeature share the implementation.

Despite the challenges of implementing the OpenFeature adapter
specifically, there are some benefits to getting something up. Doing
this adapter brings in some code for reading/writing to localstorage,
which will be shared across other adapters. Also there's also some
`SentryToolbar.init()` api changes to account for loading order and
async flag provider loading scenarios.

See Also:
-
https://openfeature.dev/specification/sections/providers/#26-provider-context-reconciliation
-
https://open-feature.github.io/js-sdk/interfaces/_openfeature_web_sdk.Client.html
-
https://github.com/open-feature/js-sdk-contrib/tree/main/libs/providers
- https://flagd.dev/providers/web/