fix: replacing organizationSlug with projectIdOrSlug where its needed (#209)
Hey team!

I noticed that some of the links that include a `?project=` actually use
the `organizationSlug` instead of the `projectIdOrSlug`. Don't know if
it's intentional, but I opened this PR anyway. Feel free to decline if
it's intentional.

I noticed this because when I clicked on the Issues panel title link to
take me to that project's issues, the project filter wasn't applied and
it showed me issues for all of my projects instead. I noticed that we're
actually sending the org slug as a project.

One more thing - I also noticed if we're using the project slug (ex.
`flashcards`) instead of the project ID (ex. `4506045517135872`) the
filter is also not applied - it only works if we use the project ID. I'm
assuming it's a change in the platform that fails to apply the filter,
but I'll open a separate issue for that.

Let me know if I can further improve my PR.