ref: Refactor InitConfig and Configuration types to simplify overrides (#240)
I wanted to fix up the types here. The InitConfig is more liberal,
that's what end-users will satisfy in when they call
`SentryToolbar.init()`.

Then the internal `Configuration` type is used with validated/hydrated
values applied to the InitConfig. Default values are applied to
`InitConfig` to turn it into a `Configuration`.