fix: Cache props for useSentryToolbar to avoid re-renders (#249)
The hook `useSentryToolbar` takes a complex set of props in the shape
(simplified): `{enabled: boolean, version: string, cdn: string,
initProps: Object}`. It's that `initProps` and it's nested object which
needs to be handled with care, because it's being passed along into the
core `init()` function and we don't want to call that function at the
wrong time.

So what I've done is put the initial `initProps` into `useRef` so that
it's cached. We'll call the core `init()` once with the first set of
`initProps` and then if the props change afterwards the core toolbar
instance won't be updated to reflect it.

Before callsites could've avoided problems by using their own `useMemo`
to ensure that the `initProps` are the same object instance each time,
this isn't needed anymore.