ref: Create static & mutable Config providers so some values can be overridden in memory (#263)
We want to be able to override stuff like `placement`, but some things
like `sentryOrigin` cannot be overridden because that would cause the
whole `<ApiProxyContextProvider>` to re-connect and be a weird
experience.

So we have the `<StaticConfigProvider>` which wraps the whole app and
nothing inside it can be changed.
Then down below we have `<MutableConfigProvider>` where the fields can
be changed. If something below here depends on a field then it will be
updated to use the new value. This means that fields like `placement`
can be updated without any problems, other fields like `debug` and
`sentryOrigin` can be updated, but nothing lower really uses them. This
is a limitation that maybe we can address in a followup iteration.

It's important is not to re-render `<ApiProxyContextProvider>` or the
other providers.