feat: Implement support for multiple placements around the screen edge (#268)
This is a foundational thing in prep for
https://github.com/getsentry/sentry-toolbar/pull/269 coming next.

With this we're adding support for the toolbar to live in 12 different
positions around the edge of the browser window.

There's 3 positions per side: for example `top-left-corner`. The major
position in this case is `top`, which dictates that the orientation will
be horizontal (icons within the toolbar flow left-to-right). The minor
position is `left` which means it'll be at the left side, in this case
left side of the top row.

When the major axis is `top` or `bottom` then the orientation will be
left-to-right. When the major is `left` or `right` then the orientation
is vertical, top-to-bottom.