feat: Implement support for drag+drop placement (#269)
This adds drag+drop support to the Toolbar. So users can drag the
toolbar into some pre-defined spots and that preference will be saved
into localStorage for them.

Below is a video of it working. And some additional visuals to show some
of the invisible stuff so you get a sense of it.

| Desc | Img |
| --- | --- |
| Video |
https://github.com/user-attachments/assets/f2d299ff-0a0e-4871-a1bf-968c7a6806f3
| All Drop Areas (usually invisible) |
![SCR-20250616-kkyw](https://github.com/user-attachments/assets/0d36121e-d797-48b9-beeb-04cb8949a08e)
| All Ghosts (usually only 1 visible during :hover) |
![SCR-20250616-kifk](https://github.com/user-attachments/assets/90139a1e-6e5a-49ed-a519-3ae2defc01a8)

Fixes #234