fix: CenterLayout.MainArea should not capture mouse events,  (#276)
The children of `<MainArea>` (which are all `<UnauthPill>` components)
should accept pointer events, but since those divs are translated down
from the actual `<MainArea>` we can remove pointer events from the main
area itself. Because of the `translate-y` css, the main area is
effectively 'empty space' on the page, used for aria correctness.