feat: Collapse any unauth dialog after 5s
This only happens if the user doesnt click to interact with the toolbar within those 5s. After the timeout, or 3 clicks outside the toolbar, it collapses down to a single icon on the right edge of the screen. Clicking this icon will bring back the toolbar in whatever its current unauth state.