feat: Collapse any unauth dialog after 5s (#277)
This only happens if the user doesnt click to interact with the toolbar
within those 5s. After the timeout, or 3 clicks outside the toolbar, it
collapses down to a single icon on the right edge of the screen.
Clicking this icon will bring back the toolbar in whatever its current
unauth state.

Here's what UnauthPill looks like when collapsed, on the right edge of
the screen (no moving it).
| State | Img |
| --- | --- |
| Normal | <img width="243" alt="SCR-20250702-mzta"
src="https://github.com/user-attachments/assets/ede556d4-5c7b-41e6-ac9f-61298f25ac34"
/>
| Hover | <img width="236" alt="Screenshot 2025-07-02 at 2 38 25 PM"
src="https://github.com/user-attachments/assets/fe3df50a-0278-4837-b5ec-da51fd1e9714"
/>


This way, if the toolbar is misbehaving and not logging people in right
away, it'll get out of the way. When an UnauthPill is expanded there's
still the dropdown and "Hide Toolbar" button that people can click to
hide it completely for this session.